--- a/components/experimental/src/dimension/provider/currency_patterns.rs
+++ b/components/experimental/src/dimension/provider/currency_patterns.rs
@@ -28,20 +28,22 @@ pub struct CurrencyPatternsDataV1<'data> {
 #[cfg(feature = "datagen")]
 impl<'data> CurrencyPatternsDataV1<'data> {
     /// Construct an instance directly from a byte slice.
-    /// 
+    ///
     /// # Safety
-    /// 
+    ///
     /// The bytes must have been returned by [`Self::as_byte_slice`].
     pub const unsafe fn from_byte_slice_unchecked(bytes: &'data [u8]) -> Self {
         Self {
             patterns: PluralPatterns {
                 strings: icu_plurals::provider::PluralElementsPackedCow {
                     elements: alloc::borrow::Cow::Borrowed(
-                        icu_plurals::provider::PluralElementsPackedULE::from_byte_slice_unchecked(bytes)
-                    )
+                        icu_plurals::provider::PluralElementsPackedULE::from_byte_slice_unchecked(
+                            bytes,
+                        ),
+                    ),
                 },
                 _phantom: core::marker::PhantomData,
-            }
+            },
         }
     }