--- a/components/experimental/src/dimension/provider/extended_currency.rs
+++ b/components/experimental/src/dimension/provider/extended_currency.rs
@@ -45,7 +45,6 @@ pub struct CurrencyExtendedDataV1<'data> {
     pub display_names: PluralElementsPackedCow<'data, str>,
 }
 
-#[cfg(feature = "datagen")]
 impl<'data> CurrencyExtendedDataV1<'data> {
     /// Construct an instance directly from a byte slice.
     ///
@@ -65,6 +64,7 @@ impl<'data> CurrencyExtendedDataV1<'data> {
     }
 
     /// Returns this instance as a byte slice.
+    #[cfg(feature = "datagen")]
     pub fn as_byte_slice(&self) -> &[u8] {
         use zerovec::ule::VarULE;
         self.display_names.elements.as_byte_slice()