--- a/components/experimental/src/dimension/provider/extended_currency.rs
+++ b/components/experimental/src/dimension/provider/extended_currency.rs
@@ -55,6 +55,10 @@ impl<'data> CurrencyExtendedDataV1<'data> {
         Self {
             display_names: icu_plurals::provider::PluralElementsPackedCow {
                 elements: alloc::borrow::Cow::Borrowed(
+                    // Safety: from_byte_slice_unchecked requires that the bytes come from a valid
+                    // slice of `PluralElementsPackedULE`, which the caller invariants enforce by
+                    // requiring callers to use as_byte_slice, which guarantees using
+                    // `PluralElementsPackedULE` in its invariants
                     icu_plurals::provider::PluralElementsPackedULE::from_byte_slice_unchecked(
                         bytes,
                     ),
@@ -64,6 +68,8 @@ impl<'data> CurrencyExtendedDataV1<'data> {
     }
 
     /// Returns this instance as a byte slice.
+    ///
+    /// Safety usable invariant: returns the slice of the internal `PluralElementsPackedULE`
     #[cfg(feature = "datagen")]
     pub fn as_byte_slice(&self) -> &[u8] {
         use zerovec::ule::VarULE;