--- a/components/experimental/src/dimension/provider/extended_currency.rs
+++ b/components/experimental/src/dimension/provider/extended_currency.rs
@@ -50,15 +50,13 @@ impl<'data> CurrencyExtendedDataV1<'data> {
     ///
     /// # Safety
     ///
-    /// The bytes must have been returned by [`Self::as_byte_slice`].
+    /// The bytes must have been returned by [`PluralElementsPackedULE::as_byte_slice`].
     pub const unsafe fn from_byte_slice_unchecked(bytes: &'data [u8]) -> Self {
         Self {
             display_names: icu_plurals::provider::PluralElementsPackedCow {
                 elements: alloc::borrow::Cow::Borrowed(
-                    // Safety: from_byte_slice_unchecked requires that the bytes come from a valid
-                    // slice of `PluralElementsPackedULE`, which the caller invariants enforce by
-                    // requiring callers to use as_byte_slice, which guarantees using
-                    // `PluralElementsPackedULE` in its invariants
+                    // Safety: this function's safety invariant guarantees that the bytes
+                    // represent a valid `PluralElementsPackedULE`
                     icu_plurals::provider::PluralElementsPackedULE::from_byte_slice_unchecked(
                         bytes,
                     ),
@@ -66,23 +64,15 @@ impl<'data> CurrencyExtendedDataV1<'data> {
             },
         }
     }
-
-    /// Returns this instance as a byte slice.
-    ///
-    /// Safety usable invariant: returns the slice of the internal `PluralElementsPackedULE`
-    #[cfg(feature = "datagen")]
-    fn as_byte_slice(&self) -> &[u8] {
-        use zerovec::ule::VarULE;
-        self.display_names.elements.as_byte_slice()
-    }
 }
 
 #[cfg(feature = "datagen")]
 impl databake::Bake for CurrencyExtendedDataV1<'_> {
     fn bake(&self, ctx: &databake::CrateEnv) -> databake::TokenStream {
+        use zerovec::ule::VarULE;
         ctx.insert("icu_experimental::dimension::provider::extended_currency");
-        let bytes = self.as_byte_slice().bake(ctx);
-        // Safety: The bytes are returned by `self.as_byte_slice()`.
+        let bytes = self.display_names.elements.as_byte_slice().bake(ctx);
+        // Safety: The bytes are returned by `PluralElementsPackedULE::as_byte_slice`.
         databake::quote! { unsafe {
             icu_experimental::dimension::provider::extended_currency::CurrencyExtendedDataV1::from_byte_slice_unchecked(#bytes)
         }}