--- a/components/experimental/src/dimension/provider/units.rs
+++ b/components/experimental/src/dimension/provider/units.rs
@@ -32,20 +32,22 @@ pub struct UnitsDisplayNameV1<'data> {
 #[cfg(feature = "datagen")]
 impl<'data> UnitsDisplayNameV1<'data> {
     /// Construct an instance directly from a byte slice.
-    /// 
+    ///
     /// # Safety
-    /// 
+    ///
     /// The bytes must have been returned by [`Self::as_byte_slice`].
     pub const unsafe fn from_byte_slice_unchecked(bytes: &'data [u8]) -> Self {
         Self {
             patterns: PluralPatterns {
                 strings: icu_plurals::provider::PluralElementsPackedCow {
                     elements: alloc::borrow::Cow::Borrowed(
-                        icu_plurals::provider::PluralElementsPackedULE::from_byte_slice_unchecked(bytes)
-                    )
+                        icu_plurals::provider::PluralElementsPackedULE::from_byte_slice_unchecked(
+                            bytes,
+                        ),
+                    ),
                 },
                 _phantom: core::marker::PhantomData,
-            }
+            },
         }
     }