--- a/components/experimental/src/dimension/provider/units.rs
+++ b/components/experimental/src/dimension/provider/units.rs
@@ -58,7 +58,7 @@ impl<'data> UnitsDisplayNameV1<'data> {
     ///
     /// Safety usable invariant: returns the slice of the internal `PluralElementsPackedULE`
     #[cfg(feature = "datagen")]
-    pub fn as_byte_slice(&self) -> &[u8] {
+    fn as_byte_slice(&self) -> &[u8] {
         use zerovec::ule::VarULE;
         self.patterns.strings.elements.as_byte_slice()
     }