--- a/components/experimental/src/dimension/provider/units.rs
+++ b/components/experimental/src/dimension/provider/units.rs
@@ -34,7 +34,8 @@ impl<'data> UnitsDisplayNameV1<'data> {
     ///
     /// # Safety
     ///
-    /// The bytes must have been returned by [`PluralElementsPackedULE::as_byte_slice`].
+    /// The bytes must have been returned by
+    /// [`icu_plurals::provider::PluralElementsPackedULE::as_byte_slice`]
     pub const unsafe fn from_byte_slice_unchecked(bytes: &'data [u8]) -> Self {
         Self {
             patterns: PluralPatterns {