… provider/data/experimental/data/currency_patterns_data_v1_marker.rs.data