Works around #5230 

(note: @sffc wrote this PR but @Manishearth opened the PR due to internet connection issues)