@sffc 2024-09-13
Before:

```
$ cargo clean; RUSTFLAGS="-Ztime-passes" /usr/bin/time -v cargo +nightly build -p icu_experimental -j1 --all-features

   Compiling icu_experimental v0.1.0 (/usr/local/google/home/sffc/projects/icu4x/components/experimental)
time:   0.000; rss:   47MB ->   48MB (   +1MB)  parse_crate
time:   0.000; rss:   48MB ->   48MB (   +0MB)  incr_comp_garbage_collect_session_directories
time:   0.000; rss:   48MB ->   49MB (   +1MB)  setup_global_ctxt
time:   0.000; rss:   51MB ->   51MB (   +0MB)  crate_injection
time:   1.021; rss:   51MB -> 1182MB (+1130MB)  expand_crate
time:   1.021; rss:   51MB -> 1182MB (+1131MB)  macro_expand_crate
time:   0.028; rss: 1182MB -> 1182MB (   +0MB)  AST_validation
time:   0.004; rss: 1182MB -> 1182MB (   +0MB)  finalize_macro_resolutions
time:   0.478; rss: 1182MB -> 1351MB ( +168MB)  late_resolve_crate
time:   0.022; rss: 1351MB -> 1351MB (   +0MB)  resolve_check_unused
time:   0.044; rss: 1351MB -> 1351MB (   +0MB)  resolve_postprocess
time:   0.550; rss: 1182MB -> 1351MB ( +169MB)  resolve_crate
time:   0.023; rss: 1271MB -> 1271MB (   +0MB)  write_dep_info
time:   0.021; rss: 1271MB -> 1272MB (   +0MB)  complete_gated_feature_checking
time:   0.077; rss: 1609MB -> 1460MB ( -149MB)  drop_ast
time:   1.621; rss: 1271MB -> 1311MB (  +39MB)  looking_for_derive_registrar
time:   1.888; rss: 1271MB -> 1307MB (  +36MB)  misc_checking_1
time:   0.271; rss: 1307MB -> 1333MB (  +26MB)  coherence_checking
time:   6.600; rss: 1307MB -> 1439MB ( +132MB)  type_check_crate
time:  35.661; rss: 1439MB -> 2239MB ( +799MB)  MIR_borrow_checking
time:   2.164; rss: 2239MB -> 2261MB (  +22MB)  MIR_effect_checking
time:   0.123; rss: 2263MB -> 2276MB (  +14MB)  module_lints
time:   0.123; rss: 2263MB -> 2276MB (  +14MB)  lint_checking
time:   0.251; rss: 2276MB -> 2276MB (   +0MB)  privacy_checking_modules
time:   0.064; rss: 2276MB -> 2270MB (   -6MB)  check_lint_expectations
time:   0.521; rss: 2261MB -> 2270MB (   +9MB)  misc_checking_3
time:   0.187; rss: 2293MB -> 2308MB (  +15MB)  monomorphization_collector_graph_walk
time:   0.033; rss: 2308MB -> 2310MB (   +2MB)  partition_and_assert_distinct_symbols
time:   0.576; rss: 2270MB -> 2292MB (  +21MB)  generate_crate_metadata
time:   3.402; rss: 2308MB -> 2327MB (  +19MB)  LLVM_passes
time:   1.049; rss: 2294MB -> 2327MB (  +33MB)  codegen_to_LLVM_IR
time:   3.492; rss: 2292MB -> 2326MB (  +34MB)  codegen_crate
time:   0.000; rss: 2326MB -> 2326MB (   +0MB)  assert_dep_graph
time:   0.345; rss: 2326MB -> 2326MB (   +0MB)  encode_query_results
time:   0.375; rss: 2326MB -> 2326MB (   +1MB)  incr_comp_serialize_result_cache
time:   0.376; rss: 2326MB -> 2326MB (   +1MB)  incr_comp_persist_result_cache
time:   0.376; rss: 2326MB -> 2326MB (   +0MB)  serialize_dep_graph
time:   0.097; rss: 2326MB -> 1308MB (-1018MB)  free_global_ctxt
time:   0.007; rss: 1308MB -> 1308MB (   +0MB)  copy_all_cgu_workproducts_to_incr_comp_cache_dir
time:   0.007; rss: 1308MB -> 1308MB (   +0MB)  finish_ongoing_codegen
time:   0.142; rss: 1282MB -> 1327MB (  +45MB)  link_rlib
time:   0.156; rss: 1282MB -> 1327MB (  +45MB)  link_binary
time:   0.159; rss: 1282MB -> 1279MB (   -3MB)  link_crate
time:   0.167; rss: 1308MB -> 1279MB (  -29MB)  link
time:  53.278; rss:   32MB ->  225MB ( +193MB)  total
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 1m 20s
        Command being timed: "cargo +nightly build -p icu_experimental -j1 --all-features"
        User time (seconds): 61.18
        System time (seconds): 17.65
        Percent of CPU this job got: 97%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 1:20.65
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 15885112
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 343
        Minor (reclaiming a frame) page faults: 6812594
        Voluntary context switches: 25417
        Involuntary context switches: 688
        Swaps: 0
        File system inputs: 0
        File system outputs: 2092016
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

After:

```
$ cargo clean; RUSTFLAGS="-Ztime-passes" /usr/bin/time -v cargo +nightly build -p icu_experimental -j1 --all-features

Compiling icu_experimental v0.1.0 (/usr/local/google/home/sffc/projects/icu4x/components/experimental)
time:   0.000; rss:   47MB ->   48MB (   +1MB)  parse_crate
time:   0.000; rss:   48MB ->   48MB (   +0MB)  incr_comp_garbage_collect_session_directories
time:   0.000; rss:   48MB ->   49MB (   +1MB)  setup_global_ctxt
time:   0.000; rss:   51MB ->   51MB (   +0MB)  crate_injection
time:   0.656; rss:   51MB ->  735MB ( +684MB)  expand_crate
time:   0.656; rss:   51MB ->  735MB ( +684MB)  macro_expand_crate
time:   0.019; rss:  735MB ->  735MB (   +0MB)  AST_validation
time:   0.270; rss:  736MB ->  843MB ( +107MB)  late_resolve_crate
time:   0.014; rss:  843MB ->  843MB (   +0MB)  resolve_check_unused
time:   0.025; rss:  843MB ->  843MB (   +0MB)  resolve_postprocess
time:   0.315; rss:  735MB ->  843MB ( +107MB)  resolve_crate
time:   0.015; rss:  801MB ->  801MB (   +0MB)  write_dep_info
time:   0.012; rss:  801MB ->  801MB (   +0MB)  complete_gated_feature_checking
time:   0.043; rss:  983MB ->  913MB (  -70MB)  drop_ast
time:   0.899; rss:  801MB ->  832MB (  +31MB)  looking_for_derive_registrar
time:   1.069; rss:  801MB ->  832MB (  +31MB)  misc_checking_1
time:   0.264; rss:  832MB ->  868MB (  +35MB)  coherence_checking
time:   3.851; rss:  832MB -> 1058MB ( +226MB)  type_check_crate
time:  18.247; rss: 1058MB -> 1468MB ( +410MB)  MIR_borrow_checking
time:   1.935; rss: 1468MB -> 1498MB (  +30MB)  MIR_effect_checking
time:   0.091; rss: 1498MB -> 1513MB (  +15MB)  module_lints
time:   0.091; rss: 1498MB -> 1513MB (  +15MB)  lint_checking
time:   0.142; rss: 1513MB -> 1513MB (   +0MB)  privacy_checking_modules
time:   0.042; rss: 1513MB -> 1513MB (   +0MB)  check_lint_expectations
time:   0.334; rss: 1498MB -> 1513MB (  +15MB)  misc_checking_3
time:   0.180; rss: 1541MB -> 1550MB (   +9MB)  monomorphization_collector_graph_walk
time:   0.033; rss: 1550MB -> 1554MB (   +5MB)  partition_and_assert_distinct_symbols
time:   0.499; rss: 1513MB -> 1538MB (  +25MB)  generate_crate_metadata
time:   3.378; rss: 1555MB -> 1575MB (  +20MB)  LLVM_passes
time:   1.049; rss: 1540MB -> 1575MB (  +35MB)  codegen_to_LLVM_IR
time:   3.466; rss: 1538MB -> 1574MB (  +36MB)  codegen_crate
time:   0.000; rss: 1573MB -> 1573MB (   +0MB)  assert_dep_graph
time:   0.266; rss: 1573MB -> 1576MB (   +3MB)  encode_query_results
time:   0.296; rss: 1573MB -> 1576MB (   +3MB)  incr_comp_serialize_result_cache
time:   0.297; rss: 1573MB -> 1576MB (   +3MB)  incr_comp_persist_result_cache
time:   0.298; rss: 1573MB -> 1576MB (   +3MB)  serialize_dep_graph
time:   0.065; rss: 1576MB -> 1038MB ( -538MB)  free_global_ctxt
time:   0.007; rss: 1038MB -> 1038MB (   +0MB)  copy_all_cgu_workproducts_to_incr_comp_cache_dir
time:   0.007; rss: 1038MB -> 1038MB (   +0MB)  finish_ongoing_codegen
time:   0.126; rss: 1011MB -> 1014MB (   +4MB)  link_rlib
time:   0.139; rss: 1011MB -> 1014MB (   +4MB)  link_binary
time:   0.141; rss: 1011MB ->  975MB (  -35MB)  link_crate
time:   0.150; rss: 1038MB ->  975MB (  -63MB)  link
time:  30.983; rss:   32MB ->  150MB ( +118MB)  total
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 57.68s
        Command being timed: "cargo +nightly build -p icu_experimental -j1 --all-features"
        User time (seconds): 47.31
        System time (seconds): 8.54
        Percent of CPU this job got: 96%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:57.71
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 3478960
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 357
        Minor (reclaiming a frame) page faults: 1386103
        Voluntary context switches: 25703
        Involuntary context switches: 489
        Swaps: 0
        File system inputs: 8040
        File system outputs: 1872008
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

File sizes:

| File | Before | After |
|---|---|---|
| units_display_name_v1_marker.rs.data | 12943294 | 6340108 |
| currency_displayname_v1_marker.rs.data | 5427952 | 5427952 |
| currency_extended_data_v1_marker.rs.data | 7337033 | 4807025 |